refactor: move model to dict conversion
